Calculating sampling locations
Made a hexagon grid to cover the area I've suggested we include in the study (approx. 25-30nm off the nearest Cornish and IOS' coastlines).
Then wrote a for loop to calculate a random location within each hexagon.  These are preliminary ideas of location, subject to change based on boat availability.